functions/requirements.txt
# Cloud Functionsで必要なPythonライブラリ

# Firebase Functions Framework
functions-framework==3.6.0
firebase-functions==0.4.3  # Firebase CLI互換性のため（デプロイ時の分析用）

# 動画処理
opencv-python==4.12.0.88
mediapipe==0.10.14

# Google Cloud Storage操作
google-cloud-storage==2.18.2

# Firestore操作（レートリミット用）
google-cloud-firestore==2.18.0

# HTTPリクエスト（Dify API、LINE API用）
requests==2.31.0

